add options to seperate files, set option bug fix, main_key_logger issue fix
